Modify ccv3 buildpack implementation to use ccv3.Client MakeRequest, MakeListRequest and MakeRequestUploadAsync methods
